requirements.txt
pytest==7.2.2
selenium==4.8.2
undetected-chromedriver==3.4.6
requests==2.26.0
pyyaml==6.0